refactor: now globally using kebab case on file names
