+++ src/components/layout/layout.tsx
@@ -1,10 +1,9 @@
-import { useEffect } from "react";
 import { useTheme } from "next-themes";
 import Head from "next/head";
-import Header from "./Header";
-import Footer from "./Footer";
+import { ReactNode, useEffect } from "react";
 import { Toaster } from "../ui/toaster";
-import { ReactNode } from "react";
+import Footer from "./footer";
+import Header from "./header";
 
 type LayoutProps = {
   children: ReactNode;
