+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
-import Layout from "@/components/layout/Layout";
+import Layout from "@/components/layout/layout";
 import { ThemeProvider } from "next-themes";
 import { NextIntlClientProvider } from "next-intl";
 import { useRouter } from "next/router";
